backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
openai
pypdf
